Streamlining the use of scripts
Ok, I know I should not put scripts in a .ejs file and include them as html, but I don't know how to do it properly yet
and honestly, it is not a priority for now, I just want to have one place where I can edit all my scripts.
I'll do it properly in the near future, I promise.