Bug solving and added feature on adding recipes
Ingredient addition id's were being sent together with the ingredient id's
when adding recipes. A small client side condition check was added to prevent
that. In the future, such check and a couple more should be done on the
server side for added security.

A new client side verification was also added to help prevent users from
forgetting to fill the recipe name and link. Such verification should be
done on the server side too, future work.